tests/wpt/web-platform-tests/trusted-types/trusted-types-duplicate-names-list.tentative.html

<!DOCTYPE html>
<head>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <meta http-equiv="Content-Security-Policy" content="trusted-types a b c*">
</head>
<body>
<script>
test(t => {
  trustedTypes.createPolicy("a", {}),
  assert_throws(new TypeError(),
                _ => trustedTypes.createPolicy("a", {}),
                "Duplicate name");

  trustedTypes.createPolicy("b", {}),

  assert_throws(new TypeError(),
                _ => trustedTypes.createPolicy("d", {}),
                "Invalid name.");
}, "TrustedTypePolicyFactory and policy list in CSP.");

</script>
</body>
